Move the SMT code to its own directory.
