+++ kore/src/Kore/Step/Condition/Evaluator.hs
@@ -9,19 +9,9 @@ Portability : portable
 -}
 module Kore.Step.Condition.Evaluator
     ( evaluate
-    , decidePredicate
     ) where
 
-import           Control.Applicative
-                 ( Alternative (..) )
-import           Control.Error
-                 ( MaybeT, runMaybeT )
-import qualified Control.Monad.Counter as Counter
-import qualified Control.Monad.State.Strict as State
-import qualified Data.Map.Strict as Map
-import           Data.Proxy
-import           Data.Reflection
-import qualified Data.Text as Text
+import Data.Reflection
 
 import           Kore.AST.Pure
 import           Kore.IndexedModule.MetadataTools
@@ -35,14 +25,11 @@ import           Kore.Step.Simplification.Data
                  ( PredicateSubstitutionSimplifier,
                  SimplificationProof (SimplificationProof), Simplifier,
                  StepPatternSimplifier (..) )
+import qualified Kore.Step.SMT.Evaluator as SmtEvaluator
 import           Kore.Step.StepperAttributes
-import           Kore.Step.TranslateSMT
 import           Kore.Unparser
 import           Kore.Variables.Fresh
                  ( FreshVariable )
-import           SMT
-                 ( MonadSMT, Result (..), SExpr (..), SMT )
-import qualified SMT
 
 {- | Attempt to evaluate a predicate.
 
@@ -82,7 +69,7 @@ evaluate
         case () of
             _ | OrOfExpandedPattern.isTrue simplified -> return (Just True)
               | OrOfExpandedPattern.isFalse simplified -> return (Just False)
-              | otherwise -> decidePredicate predicate
+              | otherwise -> SmtEvaluator.decidePredicate predicate
     let simplified' =
             case refute of
                 Just False -> ExpandedPattern.bottom
@@ -113,71 +100,3 @@ asPredicateSubstitution
             }
         , SimplificationProof
         )
-
-{- | Attempt to refute a predicate using an external SMT solver.
-
-The predicate is always sent to the external solver, even if it is trivial.
-
- -}
-decidePredicate
-    :: forall level variable m.
-       ( Given (MetadataTools level StepperAttributes)
-       , MetaOrObject level
-       , Ord (variable level)
-       , Show (variable level)
-       , Unparse (variable level)
-       , SortedVariable variable
-       , MonadSMT m
-       )
-    => Predicate level variable
-    -> m (Maybe Bool)
-decidePredicate korePredicate =
-    case isMetaOrObject (Proxy :: Proxy level) of
-        IsObject -> SMT.inNewScope $ runMaybeT $ do
-                smtPredicate <-
-                    goTranslatePredicate korePredicate
-                -- smtPredicate' <-
-                --     goTranslatePredicate (makeNotPredicate korePredicate)
-                result <- SMT.inNewScope
-                    (SMT.assert smtPredicate >> SMT.check)
-                -- result' <- SMT.inNewScope
-                --     (SMT.assert smtPredicate' >> SMT.check)
-                -- case (result, result') of
-                --     (Unsat, _) -> return False
-                --     (_, Unsat) -> return True
-                --     _ -> empty
-                case result of
-                    Unsat -> return False
-                    _ -> empty
-
-goTranslatePredicate
-    :: forall variable.
-        (Ord (variable Object)
-        , Given (MetadataTools Object StepperAttributes)
-        , Unparse (variable Object)
-        )
-    => Predicate Object variable
-    -> MaybeT SMT SExpr
-goTranslatePredicate predicate = do
-    let translator =
-            translatePredicate
-                translateUninterpreted
-                predicate
-    evalTranslator translator
-
-translateUninterpreted
-    :: Ord p
-    => SExpr  -- ^ type name
-    -> p  -- ^ uninterpreted pattern
-    -> Translator p SExpr
-translateUninterpreted t pat =
-    lookupPattern <|> freeVariable
-  where
-    lookupPattern = do
-        result <- State.gets $ Map.lookup pat
-        maybe empty (return . fst) result
-    freeVariable = do
-        n <- Counter.increment
-        var <- SMT.declare ("<" <> Text.pack (show n) <> ">") t
-        State.modify' (Map.insert pat (var, t))
-        return var
