+++ kore/test/Test/Kore/Step/Condition/Evaluator.hs
@@ -17,6 +17,7 @@ import qualified Kore.Step.Condition.Evaluator as Evaluator
 import           Kore.Step.Pattern
 import           Kore.Step.Representation.ExpandedPattern
 import           Kore.Step.Simplification.Data
+import qualified Kore.Step.SMT.Evaluator as SmtEvaluator
 import           SMT
                  ( SMT )
 import qualified SMT
@@ -108,7 +109,7 @@ div i j = mkApp intSort Builtin.tdivIntSymbol [i, j]
 assertRefuted :: CommonPredicate Object -> Assertion
 assertRefuted prop = give testMetadataTools $ do
     let expect = Just False
-    actual <- SMT.runSMT SMT.defaultConfig $ Evaluator.decidePredicate prop
+    actual <- SMT.runSMT SMT.defaultConfig $ SmtEvaluator.decidePredicate prop
     assertEqual "" expect actual
 
 unit_1 :: Assertion
